1.0.16 release and reach and hitbox fix

+++ src/main/java/keystrokesmod/module/modules/render/Nametags.java
@@ -51,8 +51,8 @@ public class Nametags extends Module {
             e.setCanceled(true);
             String str = en.getDisplayName().getFormattedText();
             if (c.isToggled()) {
-               double r = (double)(en.getHealth() / en.getMaxHealth());
-               String h = (r < 0.3D ? "§c" : (r < 0.5D ? "§6" : (r < 0.7D ? "§e" : "§a"))) + ay.round((double)en.getHealth(), 1);
+               double r = en.getHealth() / en.getMaxHealth();
+               String h = (r < 0.3D ? "§c" : (r < 0.5D ? "§6" : (r < 0.7D ? "§e" : "§a"))) + ay.round(en.getHealth(), 1);
                str = str + " " + h;
             }
 
@@ -79,10 +79,10 @@ public class Nametags extends Module {
             GlStateManager.disableTexture2D();
             if (b.isToggled()) {
                worldrenderer.begin(7, DefaultVertexFormats.POSITION_COLOR);
-               worldrenderer.pos((double)(-j - 1), (double)(-1 + i), 0.0D).color(0.0F, 0.0F, 0.0F, 0.25F).endVertex();
-               worldrenderer.pos((double)(-j - 1), (double)(8 + i), 0.0D).color(0.0F, 0.0F, 0.0F, 0.25F).endVertex();
-               worldrenderer.pos((double)(j + 1), (double)(8 + i), 0.0D).color(0.0F, 0.0F, 0.0F, 0.25F).endVertex();
-               worldrenderer.pos((double)(j + 1), (double)(-1 + i), 0.0D).color(0.0F, 0.0F, 0.0F, 0.25F).endVertex();
+               worldrenderer.pos(-j - 1, -1 + i, 0.0D).color(0.0F, 0.0F, 0.0F, 0.25F).endVertex();
+               worldrenderer.pos(-j - 1, 8 + i, 0.0D).color(0.0F, 0.0F, 0.0F, 0.25F).endVertex();
+               worldrenderer.pos(j + 1, 8 + i, 0.0D).color(0.0F, 0.0F, 0.0F, 0.25F).endVertex();
+               worldrenderer.pos(j + 1, -1 + i, 0.0D).color(0.0F, 0.0F, 0.0F, 0.25F).endVertex();
                tessellator.draw();
             }
